Fix real keyboard text input and a chain of character-generation-blocking bugs
- Add real keyboard text-input support (WM_CHAR/SDL_TEXTINPUT forwarding,
  SDL_StartTextInput) so name entry can receive real typed characters
  instead of just navigation keys.
- Extend the demo-mode input-replay harness with TYPE/BACKSPACE commands
  so name entry can be scripted and self-tested under ASAN.
- Implement Ordinal_2005 (the ARM software integer-division helper used
  ~250 places throughout uw.c) for real instead of a no-op stub; fix the
  one confirmed crash site that relied on its remainder via an
  unreliable register-leftover read.
- Implement Ordinal_1417 (CRT-style _isctype character classifier) for
  real; it was a no-op stub silently rejecting every typed character in
  the name-entry field, so no keystroke could ever be accepted.
- Fix two "dropped argument" call sites (Ordinal_1068 via FUN_00011060
  and FUN_000112a0) where the register-leftover idiom used elsewhere in
  this codebase didn't reliably survive, corrupting string-width
  measurements and causing a stack-buffer-overflow crash.
- Fix the name-entry buffer pointer: a 4-byte legacy record field was
  being used to store a full 8-byte pointer, corrupting adjacent record
  data and misfiring for unrelated records; replaced with a dedicated
  global side-table entry, matching this project's established pattern
  for pointers that don't fit their original 32-bit-sized field.
- Guard the compressed-string Huffman-style decoder (FUN_0007907c /
  FUN_00078e60 / FUN_000790e0) against a separate, deeper bug where a
  corrupted file handle caused an effectively unbounded retry loop
  (confirmed hang via lldb); now fails fast instead of freezing the
  game. Root cause of the handle corruption itself remains open.
- Quiet a noisy per-read fileio debug log line.
- Add todo.md tracking the remaining known UI/rendering issues found
  during interactive testing (missing main-menu buttons, a stray filled
  rect over the chargen stats area, missing chargen text, space key not
  producing a space in the name field, and character creation looping
  back to the first step instead of continuing into gameplay).

+++ uw.h
@@ -1457,7 +1457,7 @@ undefined4 FUN_00078d18();
 void FUN_00078e28();
 undefined1 *FUN_00078e60();
 undefined1 FUN_0007907c();
-ushort FUN_000790e0();
+int FUN_000790e0();
 undefined4 FUN_00079144();
 void FUN_0007931c();
 void FUN_00079350();
